backend/requirements.txt
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
dnspython==2.3.0
Flask==2.2.5
Flask-Cors==4.0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
pymongo==4.4.1
pyseltongue==1.0.2
six==1.16.0
typing-extensions==4.7.1
utilitybelt==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
